refactor(webapp): configure absolute path for importing files
