+++ webapp/babel.config.js
@@ -30,7 +30,17 @@ const config = {
     '@babel/plugin-syntax-dynamic-import',
     '@babel/proposal-object-rest-spread',
     '@babel/plugin-proposal-optional-chaining',
-    'babel-plugin-typescript-to-proptypes'
+    'babel-plugin-typescript-to-proptypes',
+    [
+      'module-resolver',
+      {
+        alias: {
+          src: './src',
+          '@components': './src/components',
+          '@contexts': './src/contexts'
+        }
+      }
+    ]
   ]
 }
