+++ webapp/package.json
@@ -33,6 +33,7 @@
     "@types/sdp-transform": "2.4.9",
     "@typescript-eslint/eslint-plugin": "6.21.0",
     "babel-loader": "^9.1.3",
+    "babel-plugin-module-resolver": "5.0.0",
     "core-js": "^3.36.0",
     "css-loader": "^6.10.0",
     "eslint": "8.56.0",
