+++ webapp/src/components/Conference/ParticipantList/ParticipantList.scss
