+++ webapp/src/components/Conference/Toolbar/Toolbar.scss
