+++ webapp/src/contexts/ConferenceContext/ConferenceState.ts
@@ -1,6 +1,6 @@
 import type { Client, Participant } from '@pexip/infinity'
-import type { ConnectionState } from '../../types/ConnectionState'
-import type { ConferenceConfig } from '../../types/ConferenceConfig'
+import type { ConnectionState } from 'src/types/ConnectionState'
+import type { ConferenceConfig } from 'src/types/ConferenceConfig'
 import type { Channel } from 'mattermost-redux/types/channels'
 
 interface ConferenceState {
