+++ webapp/tsconfig.json
@@ -11,6 +11,12 @@
     "resolveJsonModule": true,
     "strictNullChecks": true,
     "strictBindCallApply": true,
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "baseUrl": ".",
+    "paths": {
+      "src/*": ["./src/*"],
+      "@components/*": ["./src/components/*"],
+      "@contexts/*": ["./src/contexts/*"]
+    }
   }
 }
